+++ src/elements/Button/ButtonContent.d.ts
@@ -1,0 +1,24 @@
+import * as React from 'react';
+
+interface ContentProps {
+  [key: string]: any;
+
+  /** An element type to render as (string or function). */
+  as?: any;
+
+  /** Primary content. */
+  children?: React.ReactNode;
+
+  /** Additional classes. */
+  className?: string;
+
+  /** Initially hidden, visible on hover. */
+  hidden?: boolean;
+
+  /** Initially visible, hidden on hover. */
+  visible?: boolean;
+}
+
+const ButtonContent: React.StatelessComponent<ContentProps>;
+
+export default ButtonContent;
