+++ src/elements/Button/ButtonOr.d.ts
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+interface OrProps {
+  [key: string]: any;
+
+  /** An element type to render as (string or function). */
+  as?: any;
+
+  /** Additional classes. */
+  className?: string;
+}
+
+const ButtonOr: React.StatelessComponent<OrProps>;
+
+export default ButtonOr;
