+++ src/elements/Image/ImageGroup.d.ts
@@ -1,0 +1,22 @@
+import * as React from 'react';
+import { SemanticSIZES } from '../..';
+
+interface ImageGroupProps {
+  [key: string]: any;
+
+  /** An element type to render as (string or function). */
+  as?: any;
+
+  /** Primary content. */
+  children?: React.ReactNode;
+
+  /** Additional classes. */
+  className?: string;
+
+  /** A group of images can be formatted to have the same size. */
+  size?: SemanticSIZES;
+}
+
+const ImageGroup: React.StatelessComponent<ImageGroupProps>;
+
+export default ImageGroup;
