+++ src/elements/List/ListContent.d.ts
@@ -1,0 +1,34 @@
+import * as React from 'react';
+import { SemanticFLOATS, SemanticVERTICALALIGNMENTS } from '../..';
+
+interface ListContentProps {
+  [key: string]: any;
+
+  /** An element type to render as (string or function). */
+  as?: any;
+
+  /** Primary content. */
+  children?: React.ReactNode;
+
+  /** Additional classes. */
+  className?: string;
+
+  /** Shorthand for primary content. */
+  content?: any;
+
+  /** Shorthand for ListDescription. */
+  description?: any;
+
+  /** An list content can be floated left or right. */
+  floated?: SemanticFLOATS;
+
+  /** Shorthand for ListHeader. */
+  header?: any;
+
+  /** An element inside a list can be vertically aligned. */
+  verticalAlign?: SemanticVERTICALALIGNMENTS;
+}
+
+const ListContent: React.StatelessComponent<ListContentProps>;
+
+export default ListContent;
