+++ src/elements/List/ListHeader.d.ts
@@ -1,0 +1,21 @@
+import * as React from 'react';
+
+interface ListHeaderProps {
+  [key: string]: any;
+
+  /** An element type to render as (string or function). */
+  as?: any;
+
+  /** Primary content. */
+  children?: React.ReactNode;
+
+  /** Additional classes. */
+  className?: string;
+
+  /** Shorthand for primary content. */
+  content?: React.ReactNode;
+}
+
+const ListHeader: React.StatelessComponent<ListHeaderProps>;
+
+export default ListHeader;
