+++ src/elements/List/index.d.ts
@@ -1,200 +1,1 @@
-import * as React from 'react';
-
-import {
-  SemanticFLOATS,
-  SemanticSIZES,
-  SemanticVERTICALALIGNMENTS
-} from '../..';
-import { IconProps } from '../Icon';
-
-interface ListProps {
-  [key: string]: any;
-
-  /** An element type to render as (string or function). */
-  as?: any;
-
-  /** A list can animate to set the current item apart from the list. */
-  animated?: boolean;
-
-  /** A list can mark items with a bullet. */
-  bulleted?: boolean;
-
-  /** A list can divide its items into cells. */
-  celled?: boolean;
-
-  /** Primary content. */
-  children?: React.ReactNode;
-
-  /** Additional classes. */
-  className?: string;
-
-  /** A list can show divisions between content. */
-  divided?: boolean;
-
-  /** An list can be floated left or right. */
-  floated?: SemanticFLOATS;
-
-  /** A list can be formatted to have items appear horizontally. */
-  horizontal?: boolean;
-
-  /** A list can be inverted to appear on a dark background. */
-  inverted?: boolean;
-
-  /** Shorthand array of props for ListItem. */
-  items?: Array<any>;
-
-  /** A list can be specially formatted for navigation links. */
-  link?: boolean;
-
-  /** A list can be ordered numerically. */
-  ordered?: boolean;
-
-  /** A list can relax its padding to provide more negative space. */
-  relaxed?: boolean | 'very';
-
-  /** A selection list formats list items as possible choices. */
-  selection?: boolean;
-
-  /** A list can vary in size. */
-  size?: SemanticSIZES;
-
-  /** An element inside a list can be vertically aligned. */
-  verticalAlign?: SemanticVERTICALALIGNMENTS;
-}
-
-interface ListComponent extends React.StatelessComponent<ListProps> {
-  Content: typeof ListContent;
-  Description: typeof ListDescription;
-  Header: typeof ListHeader;
-  Icon: typeof ListIcon;
-  Item: typeof ListItem;
-  List: typeof ListList;
-}
-
-export const List: ListComponent;
-
-interface ListContentProps {
-  [key: string]: any;
-
-  /** An element type to render as (string or function). */
-  as?: any;
-
-  /** Primary content. */
-  children?: React.ReactNode;
-
-  /** Additional classes. */
-  className?: string;
-
-  /** Shorthand for primary content. */
-  content?: any;
-
-  /** Shorthand for ListDescription. */
-  description?: any;
-
-  /** An list content can be floated left or right. */
-  floated?: SemanticFLOATS;
-
-  /** Shorthand for ListHeader. */
-  header?: any;
-
-  /** An element inside a list can be vertically aligned. */
-  verticalAlign?: SemanticVERTICALALIGNMENTS;
-}
-
-export const ListContent: React.StatelessComponent<ListContentProps>;
-
-interface ListDescriptionProps {
-  [key: string]: any;
-
-  /** An element type to render as (string or function). */
-  as?: any;
-
-  /** Primary content. */
-  children?: React.ReactNode;
-
-  /** Additional classes. */
-  className?: string;
-
-  /** Shorthand for primary content. */
-  content?: React.ReactNode;
-}
-
-export const ListDescription: React.StatelessComponent<ListDescriptionProps>;
-
-interface ListHeaderProps {
-  [key: string]: any;
-
-  /** An element type to render as (string or function). */
-  as?: any;
-
-  /** Primary content. */
-  children?: React.ReactNode;
-
-  /** Additional classes. */
-  className?: string;
-
-  /** Shorthand for primary content. */
-  content?: React.ReactNode;
-}
-
-export const ListHeader: React.StatelessComponent<ListHeaderProps>;
-
-interface ListIconProps extends IconProps {
-  verticalAlign?: SemanticVERTICALALIGNMENTS;
-}
-
-export const ListIcon: React.StatelessComponent<ListIconProps>;
-
-interface ListItemProps {
-  [key: string]: any;
-
-  /** An element type to render as (string or function). */
-  as?: any;
-
-  /** A list item can active. */
-  active?: boolean;
-
-  /** Primary content. */
-  children?: React.ReactNode;
-
-  /** Additional classes. */
-  className?: string;
-
-  /** Shorthand for primary content. */
-  content?: any;
-
-  /** Shorthand for ListDescription. */
-  description?: any;
-
-  /** A list item can disabled. */
-  disabled?: boolean;
-
-  /** Shorthand for ListHeader. */
-  header?: any;
-
-  /** Shorthand for ListIcon. */
-  icon?: any;
-
-  /** Shorthand for Image. */
-  image?: any;
-
-  /** A value for an ordered list. */
-  value?: string;
-}
-
-export const ListItem: React.StatelessComponent<ListItemProps>;
-
-interface ListListProps {
-  [key: string]: any;
-
-  /** An element type to render as (string or function). */
-  as?: any;
-
-  /** Primary content. */
-  children?: React.ReactNode;
-
-  /** Additional classes. */
-  className?: string;
-}
-
-export const ListList: React.StatelessComponent<ListListProps>;
+export * from './List'
