+++ src/elements/Reveal/RevealContent.d.ts
@@ -1,0 +1,24 @@
+import * as React from 'react';
+
+interface RevealContentProps {
+  [key: string]: any;
+
+  /** An element type to render as (string or function). */
+  as?: any;
+
+  /** Primary content. */
+  children?: React.ReactNode;
+
+  /** Additional classes. */
+  className?: string;
+
+  /** A reveal may contain content that is visible before interaction. */
+  hidden?: boolean;
+
+  /** A reveal may contain content that is hidden before user interaction. */
+  visible?: boolean;
+}
+
+const RevealContent: React.StatelessComponent<RevealContentProps>;
+
+export default RevealContent;
